Add barcode-based meal logging workflow
Introduces a new add_meal blueprint with routes and templates for scanning barcodes, adding new food items, and logging meals. Updates FoodItemForm and FoodLogForm validation, changes FoodLog.amount to float, and integrates the new workflow into the daily log UI. Refactors user routes and templates to support the new meal logging process.

+++ models.py
@@ -120,13 +120,13 @@ class FoodLog(db.Model):
     )
     part_of_day = db.Column(db.Integer, nullable=False)
     user_id = db.Column(db.Integer, db.ForeignKey("user.id"), nullable=False)
-    amount = db.Column(db.Integer, nullable=False)
+    amount = db.Column(db.Float, nullable=False)
 
     def __init__(
         self,
         food_item_id: int,
         user_id: int,
-        amount: int,
+        amount: float,
         part_of_day: int,
         date_: Optional[date] = None,
     ):
